coral organizing + constants/robot container cleaning up
cleaning up coral intake/arm so that the buttons move a bit more intuitively. one key change is that the coral arm moves in sync with the elevator, so that as the elevator goes to the feeder/reef levels the coral arm will change it's angle accordingly. also cleaning up constants a bit